Support http.ResponseWriter argument to Hook()
Most Go programs that create an HTTP response do so by interacting with
an http.ResponseWriter value.  A script may now include an
http.ResponseWriter argument in its Hook() function as a way to create
an HTTP response.

This also makes it easier to use functions like http.Error(),
http.Redirect() and http.NotFound() for generating common responses.

I changed our implementation of t/hello-html.go to use this idiom.  That
way it's tested by the test suite and shown as an example in the
user documentation.